Post clean/transformation data visuals/analysis
In this R script copied from my original R markdown, I have removed all the visual printouts and file references. It includes some initial distribution investigation of prominent outcome variables and consequent cleaning. There is then some visualization of outcome variables when stratified by groupings of treatment and control (further stratified by movement or non-movement of the supplier's trader). The remainder of the file is a long stretch of regression analysis, including simple interactions, models with tow double interactions withholding the triple, and triple interaction regressions. The output of these models and intermediate steps are encoded to be presented in a stargazer table format to show the gradual effects of adding regressors on the betas and their significance levels.